Implement context manager element - New UI and supports dynamically sized label.
If the text extends past the bounds of the 'UI Context' width, we will fit the text and force it within the label. Otherwise we display it using our preferred text size and scale the label to conform to the font.

Basically if the content is too big, we scale it down. Otherwise we scale it to it's preferred width and size.